% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofm_housing_units.R
\name{ofm_county_housing_unit_data}
\alias{ofm_county_housing_unit_data}
\title{OFM Housing Unit Data by Type by County and Region}
\usage{
ofm_county_housing_unit_data(dec = 0)
}
\arguments{
\item{dec}{Number of decimals to round data - defaults to 0}
}
\value{
tibble of housing unit data by county and region from 1980 to present
}
\description{
This function pulls and cleans OFM post-censal housing unit data by county from 1980 to present
}
\examples{
housing_unit_data <- ofm_county_housing_unit_data(dec = 0)

}
