% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofm_population.R
\name{ofm_county_population_data}
\alias{ofm_county_population_data}
\title{OFM Population Data by County and Region}
\usage{
ofm_county_population_data(dec = 0)
}
\arguments{
\item{dec}{Number of decimals to round data - defaults to 0}
}
\value{
tibble of population data by county and region from 1960 to present
}
\description{
This function pulls and cleans OFM inter-censal population data by county from 1960 to 2020 and
combines it with the post-censal population data by county from 2020 onward
}
\examples{
pop_data <- ofm_county_population_data(dec = 0)

}
