% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing_near_hct.R
\name{ofm_pop_housing_near_hct}
\alias{ofm_pop_housing_near_hct}
\title{OFM Population and Housing Unit Data near High-Capacity Transit}
\usage{
ofm_pop_housing_near_hct(
  yrs = c(seq(2010, 2023, by = 1)),
  hct_title = "hct_station_area_vision_2050",
  dec = 0
)
}
\arguments{
\item{yrs}{Vector of 4 digit years to pull data for - default to 2010 through 2023}

\item{hct_title}{Name of HCT buffer in Elmer to use - deautls to VISION 2050 HCT buffers}

\item{dec}{Number of decimals to round data - defaults to 0}
}
\value{
tibble of population and housing units near HCT
}
\description{
This function pulls and cleans OFM parcelized population and housing unit data near HCT
}
\examples{
pop_housing_near_hct <- ofm_pop_housing_near_hct(yrs = c(2023))

}
